Rename DelegatingBackupRepository.
